어드민 계정 중복신청 불가 오류 해결 및 모바일 ui 임시 추가

+++ frontend/src/app/api/data/reset/route.ts
@@ -90,7 +90,7 @@ async function deletePlaylists() {
     // playlists를 받아옴 (laborCalendar)
     const playlists = await prisma.laborCalendar.findMany({
       where: {
-        month: currentMonth - 1, // 한 달 전의 데이터를 찾음
+        month: previousMonth, // 한 달 전의 데이터를 찾음
       },
       select: {
         music_url: true, // music_url 필드만 선택
@@ -100,7 +100,7 @@ async function deletePlaylists() {
     // wakeupCalendars를 받아옴 (wakeupCalendar)
     const wakeupCalendars = await prisma.wakeupCalendar.findMany({
       where: {
-        month: currentMonth - 1, // 한 달 전의 데이터를 찾음
+        month: previousMonth, // 한 달 전의 데이터를 찾음
       },
       select: {
         music_url: true, // music_url 필드만 선택
@@ -120,15 +120,18 @@ async function deletePlaylists() {
 
     // axios.delete 호출을 async로 순차적으로 처리
     for (const item of allPlaylists) {
-      await axios.delete(`${process.env.NEXTAUTH_URL}/api/deletelist`, {
-        data: {
-          playlistId: item.music_url, // playlistId를 music_url로 설정
-        },
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${process.env.CRON_SECRET}`,
-        },
-      });
+      if (item.music_url !== 'None') {
+        await axios.delete(`${process.env.NEXTAUTH_URL}/api/deletelist`, {
+          data: {
+            playlistId: item.music_url, // playlistId를 music_url로 설정
+          },
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${process.env.CRON_SECRET}`,
+          },
+        });
+  
+      }
     }
 
     return NextResponse.json({
